uwzglednienie priotytetu tagu + ukrywanie prywanych galerii

+++ src/main/java/psk/projects/dating_portal/tags/PotentialPartnerView.java
@@ -12,7 +12,7 @@ public class PotentialPartnerView {
     Long imageAvatarId;
     String description;
     String login;
-    long matchingRate;
+    double matchingRate;
 
     List<PartnerTag> tags;
 }
